# doc-files/kypo-answers-storage-swagger-open-api.yaml
---
swagger: "2.0"
info:
  version: "Version: 1.0.5"
  title: "KYPO Answers Storage - API Reference"
host: "localhost:8087"
basePath: "/kypo-rest-answers-storage/api/v1"
tags:
- name: "sandboxes"
schemes:
- "http"
- "https"
paths:
  /sandboxes:
    get:
      tags:
      - "sandboxes"
      summary: "Get all sandboxes and their answers."
      description: ""
      operationId: "findAnswersForAllSandboxes"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "The answers for all sandboxes were found."
          schema:
            $ref: "#/definitions/SandboxInfoDto"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
    post:
      tags:
      - "sandboxes"
      summary: "Store all answers for particular sandbox."
      description: ""
      operationId: "storeAnswersForParticularSandbox"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/SandboxInfoCreateDto"
      responses:
        200:
          description: "The answers for particular sandbox were created."
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
  /sandboxes/{sandboxRefId}:
    delete:
      tags:
      - "sandboxes"
      summary: "Delete sandbox_ref with all answers."
      description: ""
      operationId: "deleteSandboxWithAnswers"
      parameters:
      - name: "sandboxRefId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "The sandbox with answers was successfully deleted."
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
  /sandboxes/{sandboxRefId}/answers:
    get:
      tags:
      - "sandboxes"
      summary: "Get answers for sandbox."
      description: ""
      operationId: "findTrainingDefinitionById"
      produces:
      - "application/json"
      parameters:
      - name: "sandboxRefId"
        in: "path"
        description: "ID sandbox for that we store answers."
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "The answers for particular sandbox were found."
          schema:
            $ref: "#/definitions/SandboxInfoDto"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
  /sandboxes/{sandboxRefId}/answers/{answerIdentifier}:
    get:
      tags:
      - "sandboxes"
      summary: "Get answer for particular sandbox and by answer identifier."
      description: ""
      operationId: "findAnswerBySandboxAndIdentifier"
      produces:
      - "application/json"
      parameters:
      - name: "sandboxRefId"
        in: "path"
        description: "ID sandbox for that we store answers."
        required: true
        type: "integer"
        format: "int64"
      - name: "answerIdentifier"
        in: "path"
        description: "Identifier of the answer."
        required: true
        type: "string"
      responses:
        200:
          description: "The answer for particular sandbox and by identifier was found."
          schema:
            $ref: "#/definitions/SandboxInfoDto"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
definitions:
  ApiEntityError:
    allOf:
    - $ref: "#/definitions/ApiError"
    - type: "object"
      properties:
        entityErrorDetail:
          description: "Detail of the entity which is related to the error."
          $ref: "#/definitions/EntityErrorDetail"
      description: "A detailed error information related to the entity."
  ApiError:
    type: "object"
    properties:
      timestamp:
        type: "integer"
        format: "int64"
        example: "1574062900 (different for each type of exception)"
        description: "The time when the exception occurred"
      status:
        type: "string"
        example: "404 Not found (different for each type of exception)."
        description: "The HTTP response status code"
        enum:
        - "CONTINUE"
        - "SWITCHING_PROTOCOLS"
        - "PROCESSING"
        - "CHECKPOINT"
        - "OK"
        - "CREATED"
        - "ACCEPTED"
        - "NON_AUTHORITATIVE_INFORMATION"
        - "NO_CONTENT"
        - "RESET_CONTENT"
        - "PARTIAL_CONTENT"
        - "MULTI_STATUS"
        - "ALREADY_REPORTED"
        - "IM_USED"
        - "MULTIPLE_CHOICES"
        - "MOVED_PERMANENTLY"
        - "FOUND"
        - "MOVED_TEMPORARILY"
        - "SEE_OTHER"
        - "NOT_MODIFIED"
        - "USE_PROXY"
        - "TEMPORARY_REDIRECT"
        - "PERMANENT_REDIRECT"
        - "BAD_REQUEST"
        - "UNAUTHORIZED"
        - "PAYMENT_REQUIRED"
        - "FORBIDDEN"
        - "NOT_FOUND"
        - "METHOD_NOT_ALLOWED"
        - "NOT_ACCEPTABLE"
        - "PROXY_AUTHENTICATION_REQUIRED"
        - "REQUEST_TIMEOUT"
        - "CONFLICT"
        - "GONE"
        - "LENGTH_REQUIRED"
        - "PRECONDITION_FAILED"
        - "PAYLOAD_TOO_LARGE"
        - "REQUEST_ENTITY_TOO_LARGE"
        - "URI_TOO_LONG"
        - "REQUEST_URI_TOO_LONG"
        - "UNSUPPORTED_MEDIA_TYPE"
        - "REQUESTED_RANGE_NOT_SATISFIABLE"
        - "EXPECTATION_FAILED"
        - "I_AM_A_TEAPOT"
        - "INSUFFICIENT_SPACE_ON_RESOURCE"
        - "METHOD_FAILURE"
        - "DESTINATION_LOCKED"
        - "UNPROCESSABLE_ENTITY"
        - "LOCKED"
        - "FAILED_DEPENDENCY"
        - "UPGRADE_REQUIRED"
        - "PRECONDITION_REQUIRED"
        - "TOO_MANY_REQUESTS"
        - "REQUEST_HEADER_FIELDS_TOO_LARGE"
        - "UNAVAILABLE_FOR_LEGAL_REASONS"
        - "INTERNAL_SERVER_ERROR"
        - "NOT_IMPLEMENTED"
        - "BAD_GATEWAY"
        - "SERVICE_UNAVAILABLE"
        - "GATEWAY_TIMEOUT"
        - "HTTP_VERSION_NOT_SUPPORTED"
        - "VARIANT_ALSO_NEGOTIATES"
        - "INSUFFICIENT_STORAGE"
        - "LOOP_DETECTED"
        - "BANDWIDTH_LIMIT_EXCEEDED"
        - "NOT_EXTENDED"
        - "NETWORK_AUTHENTICATION_REQUIRED"
      message:
        type: "string"
        example: "The IDMGroup could not be found in database (different for each\
          \ type of exception)."
        description: "The specific description of the ApiError."
      errors:
        type: "array"
        example: "[The requested resource was not found (different for each type of\
          \ exception).]"
        description: "The list of main reasons of the ApiError."
        items:
          type: "string"
      path:
        type: "string"
        example: "/kypo2-rest-user-and-group/api/v1/groups/1000 (different for each\
          \ type of exception)."
        description: "The requested URI path which caused error."
    description: "Superclass for classes ApiEntityError and ApiMicroserviceError"
  EntityErrorDetail:
    type: "object"
    properties:
      entity:
        type: "string"
        example: "IDMGroup"
        description: "Class of the entity."
      identifier:
        type: "string"
        example: "id"
        description: "Identifier of the entity."
      identifierValue:
        type: "object"
        example: "1"
        description: "Value of the identifier."
      reason:
        type: "string"
        example: "Group with same name already exists."
        description: "Detailed message of the exception"
  SandboxAnswersCreateDto:
    type: "object"
    properties:
      answerContent:
        type: "string"
        example: "nmap 192.168.0.1"
        description: "The content of the KYPO answer in particular (phase/level)"
      answerIdentifier:
        type: "string"
        example: "sandbox-1-2-answer"
        description: "The identifier of the KYPO answer"
  SandboxAnswersDto:
    type: "object"
    properties:
      answerContent:
        type: "string"
        example: "nmap 192.168.0.1"
        description: "The content of the KYPO answer in particular (phase/level)"
      answerIdentifier:
        type: "string"
        example: "sandbox-1-2-answer"
        description: "The identifier of the KYPO answer"
  SandboxInfoCreateDto:
    type: "object"
    properties:
      sandboxRefId:
        type: "integer"
        format: "int64"
        example: 12
        description: "The identifier of a sandbox for that we store the answers"
      sandboxAnswers:
        type: "array"
        description: "The answers for given sandbox"
        items:
          $ref: "#/definitions/SandboxAnswersCreateDto"
  SandboxInfoDto:
    type: "object"
    properties:
      sandboxRefId:
        type: "integer"
        format: "int64"
        example: 12
        description: "The identifier of a sandbox for that we store the answers"
      sandboxAnswers:
        type: "array"
        description: "The answers for given sandbox"
        items:
          $ref: "#/definitions/SandboxAnswersDto"